+++ lib/src/relations/relation.dart
@@ -2,7 +2,7 @@ import 'package:sqfly/sqfly.dart';
 import 'package:sqfly/src/dao.dart';
 
 abstract class Relation<T extends Dao> {
-  Dao get dao => Sqfly.daos[T];
+  Dao get dao => Sqfly.instance.daos[T];
 
   const Relation();
 }