+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.0.0-alpha.2
+
+- Use `initialize` instead of `Sqfly().init()`.
+- Add ability to pass custom database path.
+- Use name property without prefixing.
+
 # 1.0.0-alpha.1
 
 - First release.